+++ outputs.tf
@@ -53,6 +53,11 @@ output "application_gateway_id" {
   value       = module.appgw.application_gateway_id
 }
 
+output "application_gateway_identity_principal_id" {
+  description = "Id of the managed service identity of the application gateway used by AKS"
+  value       = var.appgw_identity_enabled ? azurerm_user_assigned_identity.appgw_assigned_identity[0].principal_id : null
+}
+
 output "application_gateway_name" {
   description = "Name of the application gateway used by AKS"
   value       = module.appgw.application_gateway_name