+++ providers.tf
@@ -14,6 +14,5 @@ provider "helm" {
     client_certificate     = base64decode(azurerm_kubernetes_cluster.aks.kube_config[0].client_certificate)
     client_key             = base64decode(azurerm_kubernetes_cluster.aks.kube_config[0].client_key)
     cluster_ca_certificate = base64decode(azurerm_kubernetes_cluster.aks.kube_config[0].cluster_ca_certificate)
-
   }
 }