+++ tools/cert-manager/outputs.tf
@@ -1,4 +1,4 @@
 output "namespace" {
   description = "Namespace used for Cert Manager"
-  value       = try(kubernetes_namespace.cert-manager.0.metadata.0.name, "")
+  value       = try(kubernetes_namespace.cert_manager[0].metadata[0].name, "")
 }