+++ tools/cert-manager/r-cert-manager.tf
@@ -1,4 +1,4 @@
-resource "kubernetes_namespace" "cert-manager" {
+resource "kubernetes_namespace" "cert_manager" {
   count = var.enable_cert_manager ? 1 : 0
   metadata {
     name = var.cert_manager_namespace
@@ -8,12 +8,12 @@ resource "kubernetes_namespace" "cert-manager" {
   }
 }
 
-resource "helm_release" "cert-manager" {
+resource "helm_release" "cert_manager" {
   count      = var.enable_cert_manager ? 1 : 0
   name       = "cert-manager"
   chart      = "cert-manager"
   repository = var.cert_manager_chart_repository
-  namespace  = kubernetes_namespace.cert-manager.0.metadata.0.name
+  namespace  = kubernetes_namespace.cert_manager[0].metadata[0].name
   version    = var.cert_manager_chart_version
   dynamic "set" {
     for_each = local.cert_manager_values