+++ tools/kured/versions.tf
@@ -3,11 +3,7 @@ terraform {
   required_providers {
     helm = {
       source  = "hashicorp/helm"
-      version = "1.1.1"
-    }
-    kubernetes = {
-      source  = "hashicorp/kubernetes"
-      version = ">= 1.11.1"
+      version = ">= 2.3.0"
     }
   }
 }