+++ variables.tf
@@ -70,6 +70,75 @@ variable "enable_pod_security_policy" {
   default     = false
 }
 
+variable "private_cluster_enabled" {
+  description = "Configure AKS as a Private Cluster : https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/kubernetes_cluster#private_cluster_enabled"
+  type        = bool
+  default     = false
+}
+
+variable "vnet_id" {
+  description = "Vnet id that Aks MSI should be network contributor in a private cluster"
+  type        = string
+  default     = null
+}
+
+variable "appgw_identity_enabled" {
+  description = "Configure a managed service identity for Application gateway used with AGIC (useful to configure ssl cert into appgw from keyvault)"
+  type        = bool
+  default     = false
+}
+
+variable "private_dns_zone_type" {
+  type        = string
+  default     = "System"
+  description = <<EOD
+Set AKS private dns zone if needed and if private cluster is enabled (privatelink.<region>.azmk8s.io)
+- "Custom" : You will have to deploy a private Dns Zone on your own and pass the id with <private_dns_zone_id> variable
+If this settings is used, aks user assigned identity will be "userassigned" instead of "systemassigned"
+and the aks user must have "Private DNS Zone Contributor" role on the private DNS Zone
+- "System" : AKS will manage the private zone and create it in the same resource group as the Node Resource Group
+- "None" : In case of None you will need to bring your own DNS server and set up resolving, otherwise cluster will have issues after provisioning.
+
+https://registry.terraform.io/providers/hashicorp/azurerm/latest/docs/resources/kubernetes_cluster#private_dns_zone_id
+EOD
+}
+
+variable "private_dns_zone_id" {
+  type        = string
+  default     = null
+  description = "Id of the private DNS Zone when <private_dns_zone_type> is custom"
+}
+
+variable "aks_user_assigned_identity_resource_group_name" {
+  description = "Resource Group where to deploy the aks user assigned identity resource. Used when private cluster is enabled and when Azure private dns zone is not managed by aks"
+  type        = string
+  default     = null
+}
+
+variable "aks_user_assigned_identity_custom_name" {
+  description = "Custom name for the aks user assigned identity resource"
+  type        = string
+  default     = null
+}
+
+variable "aks_sku_tier" {
+  description = "aks sku tier. Possible values are Free ou Paid"
+  type        = string
+  default     = "Free"
+}
+
+variable "appgw_user_assigned_identity_custom_name" {
+  description = "Custom name for the application gateway user assigned identity resource"
+  type        = string
+  default     = null
+}
+
+variable "appgw_user_assigned_identity_resource_group_name" {
+  description = "Resource Group where to deploy the application gateway user assigned identity resource"
+  type        = string
+  default     = null
+}
+
 variable "default_node_pool" {
   description = <<EOD
 Default node pool configuration:
@@ -92,7 +161,6 @@ map(object({
     enable_node_public_ip = bool
 }))
 ```
-
 EOD
 
   type    = map(any)
@@ -152,6 +220,12 @@ variable "nodes_pools" {
 
 }
 
+variable "container_registries_id" {
+  description = "List of Azure Container Registries ids where AKS needs pull access."
+  type        = list(string)
+  default     = []
+}
+
 variable "diagnostic_settings_custom_name" {
   description = "Custom name for Azure Diagnostics for AKS."
   type        = string
@@ -186,7 +260,7 @@ variable "diagnostic_settings_metric_categories" {
 # AGIC variables
 ##########################
 
-variable "enable_agic" {
+variable "agic_enabled" {
   description = "Enable Application gateway ingress controller"
   type        = bool
   default     = true
@@ -232,6 +306,12 @@ variable "appgw_settings" {
   description = "Application gateway configuration settings. Default dummy configuration"
   type        = map(any)
   default     = {}
+}
+
+variable "appgw_ssl_certificates_configs" {
+  description = "Application gateway ssl certificates configuration"
+  type        = list(map(string))
+  default     = []
 }
 
 ##########################
@@ -332,6 +412,7 @@ map(object({
   account_replication_type = string 
   tags                     = map(any) 
   allowed_cidrs            = list(string) 
+  allowed_subnet_ids       = list(string) 
   container_name           = string 
 }))
 ```
@@ -450,3 +531,5 @@ variable "appgw_private_ip" {
   type        = string
   default     = null
 }
+
+