+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # v4.3.0 - unreleased
 
+Breaking
+  * AZ-485: Add AKS Private Cluster Feature
+  
+Added
+  * AZ-485: Add Msi Identity for Applicationg Gateway with Agic
+  * AZ-485: Add Private DNS Zone support
+  * AZ-485: Add Azure Container Registry Permissions
+  * AZ-485: Allow to configure AKS SKU
+
 Changed
   * AZ-532: Revamp README with latest `terraform-docs` tool
   * AZ-530: Cleanup module, fix linter errors
@@ -64,4 +73,3 @@ Fixed
 
 Added
   * AZ-123: First version of the AKS module
-